+++ NeoAnalysis_Py3.5/NeoAnalysis/TransFile.py
@@ -101,7 +101,7 @@ class transfile():
                 times = np.array(ang.times.rescale(pq.ms))
                 data = np.array(ang.rescale(pq.mV))
                 data.shape = data.shape[0]
-                sampling_rate = float(ang.sampling_rate.rescale(pq.Hz))
+                sampling_rate = int(ang.sampling_rate.rescale(pq.Hz))
                 analogs['analogs'][key] = {'data':data, 'start_time':times[0], 'sampling_rate':sampling_rate}
         
         # event entity
@@ -144,7 +144,7 @@ class transfile():
                 times = np.array(ang.times.rescale(pq.ms))
                 data = np.array(ang.rescale(pq.mV))
                 data.shape = data.shape[0]
-                sampling_rate = float(ang.sampling_rate.rescale(pq.Hz))
+                sampling_rate = int(ang.sampling_rate.rescale(pq.Hz))
                 analogs['analogs'][key] = {'data':data,'start_time':times[0],'sampling_rate':sampling_rate}
 
         # event entity
@@ -186,7 +186,7 @@ class transfile():
                 times = np.array(ang.times.rescale(pq.ms))
                 data = np.array(ang.rescale(pq.mV))
                 data.shape = data.shape[0]
-                sampling_rate = float(ang.sampling_rate.rescale(pq.Hz))
+                sampling_rate = int(ang.sampling_rate.rescale(pq.Hz))
                 analogs['analogs'][key] = {'data':data, 'start_time':times[0], 'sampling_rate':sampling_rate}
 
         # event entity
@@ -226,7 +226,7 @@ class transfile():
                 times = np.array(ang.times.rescale(pq.ms))
                 data = np.array(ang.rescale(pq.mV))
                 data.shape = data.shape[0]
-                sampling_rate = float(ang.sampling_rate.rescale(pq.Hz))
+                sampling_rate = int(ang.sampling_rate.rescale(pq.Hz))
                 analogs['analogs'][key] = {'data':data, 'start_time':times[0], 'sampling_rate':sampling_rate}
 
         # event entity
@@ -270,7 +270,7 @@ class transfile():
                 times = np.array(ang.times.rescale(pq.ms))
                 data = np.array(ang.rescale(pq.mV))
                 data.shape = data.shape[0]
-                sampling_rate = float(ang.sampling_rate.rescale(pq.Hz))
+                sampling_rate = int(ang.sampling_rate.rescale(pq.Hz))
                 analogs['analogs'][key] = {'data':data, 'start_time':times[0], 'sampling_rate':sampling_rate}
 
         # event entity
